TheTilt/requirements.txt
pandas==2.2.1
pyarrow==15.0.0
lightgbm==4.3.0
scikit-learn==1.4.1.post1
requests==2.31.0
pyyaml==6.0.1
flask==3.0.0
Flask-Cors==4.0.0
